Move class name into constant
