+++ platform-sdk/consensus-otter-tests/src/test/java/org/hiero/otter/fixtures/InstrumentedNodeTest.java
@@ -1,16 +1,13 @@
 // SPDX-License-Identifier: Apache-2.0
-package org.hiero.otter.fixtures.turtle;
+package org.hiero.otter.fixtures;
 
 import static org.hiero.otter.fixtures.OtterAssertions.assertThat;
 
 import edu.umd.cs.findbugs.annotations.NonNull;
 import java.time.Duration;
 import java.util.stream.Stream;
-import org.hiero.otter.fixtures.InstrumentedNode;
-import org.hiero.otter.fixtures.Network;
-import org.hiero.otter.fixtures.TestEnvironment;
-import org.hiero.otter.fixtures.TimeManager;
 import org.hiero.otter.fixtures.container.ContainerTestEnvironment;
+import org.hiero.otter.fixtures.turtle.TurtleTestEnvironment;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.MethodSource;
