+++ platform-sdk/consensus-otter-tests/src/testFixtures/java/org/hiero/otter/fixtures/container/InstrumentedContainerNode.java
@@ -1,6 +1,8 @@
 // SPDX-License-Identifier: Apache-2.0
 package org.hiero.otter.fixtures.container;
 
+import static org.hiero.otter.fixtures.internal.InstrumentedClasses.INSTRUMENTED_EVENT_CREATOR;
+
 import com.swirlds.platform.config.ModuleConfig_;
 import edu.umd.cs.findbugs.annotations.NonNull;
 import java.nio.file.Path;
@@ -36,10 +38,7 @@ public class InstrumentedContainerNode extends ContainerNode implements Instrume
             @NonNull final ImageFromDockerfile dockerImage,
             @NonNull final Path outputDirectory) {
         super(selfId, keysAndCerts, network, dockerImage, outputDirectory);
-        configuration()
-                .set(
-                        ModuleConfig_.EVENT_CREATOR_MODULE,
-                        "org.hiero.consensus.event.creator.instrumented.InstrumentedEventCreator");
+        configuration().set(ModuleConfig_.EVENT_CREATOR_MODULE, INSTRUMENTED_EVENT_CREATOR);
     }
 
     /**
