+++ platform-sdk/consensus-otter-tests/src/testFixtures/java/org/hiero/otter/fixtures/internal/InstrumentedClasses.java
@@ -1,0 +1,12 @@
+// SPDX-License-Identifier: Apache-2.0
+package org.hiero.otter.fixtures.internal;
+
+/**
+ * A utility class that holds the fully qualified class names of the entry points to our instrumented modules.
+ */
+public class InstrumentedClasses {
+
+    /** The fully qualified class name of the instrumented event creator. */
+    public static final String INSTRUMENTED_EVENT_CREATOR =
+            "org.hiero.consensus.event.creator.instrumented.InstrumentedEventCreator";
+}
